Lucky v3
I cleaned up the formatting, and I reverted a few things:

measure 8 - This chord progression was deliberate. I think starting and
ending on minor sounds weaker; there's less feeling of movement. I also
prefer the downward motion in the bass parts, or at least the bottom
one, contrasted with the sops being up high. It makes it more of a
surprise when the altos & sops jump in. So I changed some things back,
but I tried to keep the improved voicing.

all choruses - These chord progressions were also deliberately different from
the song, because the upper solo part often didn't fit into the original
chords, and I think(?) that might be a bigger problem a cappella than in the
original song. The main problem is the new "F" chord, which is spelled FGABC,
all right next to each other. Of course the solo parts resolve down right
after, but as they keep descending they land on all sorts of weird dissonance
on the way down. Also, I like the old sop melody. I had sort of written
everything around that sop part and solo parts.

measure 33 - The slightly ambiguous chord here was deliberate. My
intention for the bridge is too feel a bit off-balance, in contrast to
the strong, straightforward sound of the chorus. That's why I saved the
clear G chord for later in the bridge, because I didn't want it to
completely resolve until later.

measure 96 - This chord was, like, extra deliberate. I think repeating
the same, consonant chords sounds bland. (Note the fermata I had put in
to emphasize the difference. If we were to go with Am, the fermata
wouldn't make sense.) If anything, I was considering putting in more
dissonance. If the concern is that F2 is too low for all the basses, we
could move the F3 from tenor to bass, although that gives the upper
bases have a 7th jump afterwards.